Refactor components to remove feature flags and streamline routing
- Eliminated the use of feature flags related to URL routing across various components, including Header, Body, and ProgressOverview.
- Updated the routing logic to always enable URL routing, simplifying the navigation structure.
- Removed legacy navigation code and associated tests, ensuring a cleaner and more maintainable codebase.
- Introduced new button and toggle variants for consistent styling across the UI components.
