+++ src/features/layout/components/Header.tsx
@@ -5,7 +5,6 @@ import MiscellaneousIcons from '@/icons/miscellaneous';
 import { Button } from '@/components/ui/button';
 import { ThemeContext } from '@/context/ThemeContext';
 import { useHabitsContext } from '@/features/habits/hooks/useHabitsContext';
-import { featureFlags } from '@/config/featureFlags';
 import { trackNavigationClick } from '@/utils/monitoring';
 const { CheckCircle2, Edit, Moon, Plus, Sun, BarChart2, Calendar } =
   MiscellaneousIcons;
@@ -32,97 +31,55 @@ export default function Header() {
             aria-label="User actions"
             className="bg-opacity-20 hover:bg-opacity-30 flex space-x-2 rounded-full p-1 transition-all duration-300 sm:space-x-3 sm:p-2"
           >
-            {featureFlags.isUrlRoutingEnabled ? (
-              <>
-                <NavLink
-                  to="/"
-                  onClick={trackNavigationClick}
-                  className={({ isActive }) =>
-                    `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
-                      isActive
-                        ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
-                        : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
-                    }`
-                  }
-                  aria-label="Daily Tracker"
-                >
-                  <Calendar
-                    aria-hidden="true"
-                    className="h-4 w-4 sm:h-6 sm:w-6"
-                  />
-                </NavLink>
-                <NavLink
-                  to="/habits"
-                  onClick={trackNavigationClick}
-                  className={({ isActive }) =>
-                    `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
-                      isActive
-                        ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
-                        : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
-                    }`
-                  }
-                  aria-label="Habits Overview"
-                >
-                  <Edit aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
-                </NavLink>
-                <NavLink
-                  to="/progress"
-                  onClick={trackNavigationClick}
-                  className={({ isActive }) =>
-                    `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
-                      isActive
-                        ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
-                        : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
-                    }`
-                  }
-                  aria-label="Progress Overview"
-                >
-                  <BarChart2
-                    aria-hidden="true"
-                    className="h-4 w-4 sm:h-6 sm:w-6"
-                  />
-                </NavLink>
-                <Button
-                  className="flex h-8 w-8 items-center justify-center rounded-full bg-(--color-brand-primary)/20 p-0 text-(--color-brand-primary) shadow-md transition-all duration-300 hover:bg-(--color-brand-secondary)/30 hover:shadow-lg sm:h-10 sm:w-10 dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20 [&_svg:not([class*='size-'])]:size-4 sm:[&_svg:not([class*='size-'])]:size-6"
-                  onClick={toggleShowAddHabitDialog}
-                  aria-label="Add Habit"
-                >
-                  <Plus aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
-                </Button>
-              </>
-            ) : (
-              <>
-                {/* Legacy navigation buttons - these won't function but maintain visual compatibility */}
-                <Button
-                  className="h-8 w-8 rounded-full bg-(--color-brand-primary) p-0 shadow-md transition-all duration-300 hover:bg-(--color-brand-secondary) hover:shadow-lg sm:h-10 sm:w-10 dark:bg-(--color-surface-secondary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20"
-                  onClick={() =>
-                    console.warn(
-                      'Legacy navigation mode - enable URL routing for navigation',
-                    )
-                  }
-                  aria-label="Habits (Legacy Mode)"
-                >
-                  <Edit
-                    aria-hidden="true"
-                    className="h-4 w-4 text-white sm:h-6 sm:w-6 dark:text-(--color-brand-primary)"
-                  />
-                </Button>
-                <Button
-                  className="h-8 w-8 rounded-full bg-(--color-brand-primary) p-0 shadow-md transition-all duration-300 hover:bg-(--color-brand-secondary) hover:shadow-lg sm:h-10 sm:w-10 dark:bg-(--color-surface-secondary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20"
-                  aria-label="Progress (Legacy Mode)"
-                  onClick={() =>
-                    console.warn(
-                      'Legacy navigation mode - enable URL routing for navigation',
-                    )
-                  }
-                >
-                  <BarChart2
-                    aria-hidden="true"
-                    className="h-4 w-4 text-white sm:h-6 sm:w-6 dark:text-(--color-brand-primary)"
-                  />
-                </Button>
-              </>
-            )}
+            <NavLink
+              to="/"
+              onClick={trackNavigationClick}
+              className={({ isActive }) =>
+                `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
+                  isActive
+                    ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
+                    : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
+                }`
+              }
+              aria-label="Daily Tracker"
+            >
+              <Calendar aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
+            </NavLink>
+            <NavLink
+              to="/habits"
+              onClick={trackNavigationClick}
+              className={({ isActive }) =>
+                `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
+                  isActive
+                    ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
+                    : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
+                }`
+              }
+              aria-label="Habits Overview"
+            >
+              <Edit aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
+            </NavLink>
+            <NavLink
+              to="/progress"
+              onClick={trackNavigationClick}
+              className={({ isActive }) =>
+                `flex h-8 w-8 items-center justify-center rounded-full p-0 shadow-md transition-all duration-300 sm:h-10 sm:w-10 ${
+                  isActive
+                    ? 'bg-(--color-brand-secondary) text-white shadow-lg hover:bg-(--color-brand-tertiary) dark:bg-(--color-brand-primary) dark:shadow-(--color-brand-secondary)/20 dark:hover:bg-(--color-brand-secondary)'
+                    : 'bg-(--color-brand-primary)/20 text-(--color-brand-primary) hover:bg-(--color-brand-secondary)/30 hover:shadow-lg dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20'
+                }`
+              }
+              aria-label="Progress Overview"
+            >
+              <BarChart2 aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
+            </NavLink>
+            <Button
+              className="flex h-8 w-8 items-center justify-center rounded-full bg-(--color-brand-primary)/20 p-0 text-(--color-brand-primary) shadow-md transition-all duration-300 hover:bg-(--color-brand-secondary)/30 hover:shadow-lg sm:h-10 sm:w-10 dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20 [&_svg:not([class*='size-'])]:size-4 sm:[&_svg:not([class*='size-'])]:size-6"
+              onClick={toggleShowAddHabitDialog}
+              aria-label="Add Habit"
+            >
+              <Plus aria-hidden="true" className="h-4 w-4 sm:h-6 sm:w-6" />
+            </Button>
 
             <Button
               className="flex h-8 w-8 items-center justify-center rounded-full bg-(--color-brand-primary)/20 p-0 text-(--color-brand-primary) shadow-md transition-all duration-300 hover:bg-(--color-brand-secondary)/30 hover:shadow-lg sm:h-10 sm:w-10 dark:bg-(--color-surface-secondary) dark:text-(--color-brand-primary) dark:hover:bg-(--color-surface-tertiary) dark:hover:shadow-(--color-brand-primary)/20 [&_svg:not([class*='size-'])]:size-4 sm:[&_svg:not([class*='size-'])]:size-6"
